+++ CHANGELOG.md
@@ -15,6 +15,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Removed
+- Renamed internal classes (removed superfluous `Diagnostics`-prefix)
+
+## Added
+- Ability to configure the formatting using a global configuration method (`llvm_diagnostics.config(...)`)
+
 ## [2.1.2] - 2022-05-03
 ### Added
 - Documentation (incl. Contribution, Security, Maintenance)